make watcher age a long, since times are

+++ src/main/scala/committedtocache.scala
@@ -5,7 +5,7 @@ import akka.event.Logging
 import java.io.File
 
 class Watcher(
-  src: File, min_age: Int, cache: S3BlobCache, s3: S3) extends Actor {
+  src: File, min_age: Long, cache: S3BlobCache, s3: S3) extends Actor {
 
   val mover = context.actorOf(
     Props(classOf[MoveActor], cache, s3).withRouter(
